CI: distribution-version: handle minor version numbers
We have always planned to support minor version numbers, but did not
implement support for them in the checker script yet.

This change allows stable version numbers like these:

  v24.09
  v24.09.1
  v24.09.123

And devel version numbers like this one:

  v24.09+dev

But not the combination:

  v24.09.1+dev
  v24.09.123+dev

Because "+dev" should only be used on the current development branch,
while minor numbers should only be used on stable branches.
There is no branch on which both may be used.

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>